+++ xpilot/src/client/guimap.h
@@ -25,7 +25,7 @@
 #define GUIMAP_H
 
 void Gui_paint_cannon(int x, int y, int type);
-void Gui_paint_fuel(int x, int y, int fuel);
+void Gui_paint_fuel(int x, int y, double fuel);
 void Gui_paint_base(int x, int y, int id, int team, int type);
 void Gui_paint_decor(int x, int y, int xi, int yi, int type,
 		     bool last, bool more_y);
